[core][wip] Created an abstract of residual terms
They are three main important point in this commit:

1. In general, we are instered in define residual functions either
depends linerly on the state (or control) or with a non-linear relation
between the states (or control). This commit provides both due to
efficient computation.

2. I created a cost manager class that containts both terminal and
running cost functions. This provides an unified description of the cost
functions without loosing the capabilities of parallelizing the
computation of it.

3. I included a quadratic cost abstract class which can be used for any
kind of residual term. This is equivalent to a least state-(control)
cost function.

All the classes have been well documented. Furthermore, I put all the
data-structure class in one single file (i.e. data.py) for easier
tracking.