[core] Refactored the argument order of the dynamics methods
Our convention is to pass first the data variable, as in the cost-
related classes. Additionally, I improved the documentation of the
dynamics abstract class and its data structure. Finally, note that
abstract classes create a essential data, however, the user can re-
implement the createData method for special cases.