[core] State difference
During the forward pass we need to compute the error between the new and
nominal states, however, this isn't computed in general by the Euclidean
difference. Nonetheless, we could want to compute the difference of
SE(3) groups (i.e. floating-base positions). For that we need to
specific a custom code that depends on the system's dynamic model.