[cost][bugfix] Created the data per each single cost function
We were using internal data of the cost model class. Since we have one
copy per each cost model, then we have one copy of the internal cost
data. Instead we should have N copy of that data, where N is the number
of knots.

In practice, we are using the last written value of these internal data.
This was producing a wrongly behaviour of the DDP solver.