[unittest] Comment format " threshold value was XX, is now YY (see assertNumDiff.__doc__)" to test_dynamics_derivatives
The previous values for ABA/RNEA derivatives were changed
ine29443e921fd68bdb514. And the rest of the values were changed in
8d8df6febf942ade957ed60f5e9f7905.

An important observation is that the new tolerance values are lower
(with the exception of dcid_du, than before changing the disturbance
values. As I observed preliminary, the new disturbance value produces
much better results, and it can be the explanation why the previous
tolerances were high.