[logs] These changes are due to the imposition of Vxx symmetric
In previous commit (095cbe4f6689afc1af019066), there is added a
numerical trick in order to impose symmetric in Vxx. With long number of
action, numerical rounding produces unsymmetric matrix and the solver
struggles to find a solution.

It's difficult to quantify how much this numerical trick is affecting
the ground-truth values. Please note that previous values were reported
with a code that was generated non-symmetric Vxx matrices.