[cost] Remove with_residuals since it is redundant or not useful
For each cost, we need to compute the residuals and store its values. As
you can see, we were always using the with_residuals_ = true.

For the cost sum, the computation of the residuals are not fair or
useful. For instance, they cannot be used to compute the second order
derivatives based on Gauss approximation. The reason is that they do not
encode the partial derivatives of the activation model.