[solver] Changed stopping criteria to |dV[0]|, instead of ||Qu||
dV[0] represents the gradient of the cost function for the entire
trajectory. Instead ||Qu|| assumes that the algoritm is pure
single-shooting approach that depends in the control only.

In the log files (i.e., double pendulum, quadrotor, and arm
manipulation, we can also observe that the ||Qu|| is stopping much
earlier than desired the algorithm. It doesn't consider the further
improvement needed for changes in xs.